No need to create same listener in both listener classes
