+++ src/main/java/com/civservers/plugins/simplecommandblocker/Listeners1131.java
@@ -18,38 +18,6 @@ public class Listeners1131 implements Listener {
 	public Listeners1131(SimpleCommandBlocker plugin) {
 		this.plugin = plugin;
 	}
-	
-    
-    @EventHandler
-    public void onCommand(PlayerCommandPreprocessEvent e) {
-    	Player player = e.getPlayer();
-    	String p_uuid = player.getUniqueId().toString();
-    	List<String> trustList = plugin.getConfig().getStringList("trustlist");
-    	
-    	if (!player.isOp() && !player.hasPermission("scb.bypass") && !player.hasPermission("simplecommandblocker.bypass") && !trustList.contains(p_uuid)) {
-	    	if (plugin.getConfig().getBoolean("blockCommands")) {
-	    		String cmd = "";
-		    	List<String> allowedCommands = plugin.getConfig().getStringList("allowed_commands");
-		    	if (e.getMessage().indexOf(" ") >= 0) {
-		    		cmd = e.getMessage().substring(1, e.getMessage().indexOf(" "));
-		    	} else {
-		    		cmd = e.getMessage().substring(1, e.getMessage().length());
-		    	}
-		    	
-		    	if (!allowedCommands.contains(cmd)) {
-		    		Utilities.sendPlayer(player, ChatColor.RED + plugin.getMessages().get("no_cmds").toString().replaceAll("<cmd>", cmd));
-		    		if (plugin.getConfig().getBoolean("play_sound"))
-			    		try {
-			    			player.playSound(player.getLocation(), Sound.valueOf(plugin.getConfig().getString("sound").toString()), 2, 1);
-			    		}
-			    		catch (Exception err) {
-			    			Utilities.sendConsole(ChatColor.RED + plugin.getMessages().get("bad_sound_config").toString());
-			    		}
-		    		e.setCancelled(true);
-		    	}
-	    	}
-    	}
-    }
         
    @EventHandler
    public void onPlayerTab(PlayerCommandSendEvent e) {
