+++ src/main/java/com/civservers/plugins/simplecommandblocker/SimpleCommandBlocker.java
@@ -6,7 +6,6 @@ package com.civservers.plugins.simplecommandblocker;
 import java.util.Map;
 
 import org.bukkit.Bukkit;
-import org.bukkit.event.Listener;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import net.md_5.bungee.api.ChatColor;
@@ -30,12 +29,13 @@ public final class SimpleCommandBlocker extends JavaPlugin {
 		if (mcVer.contains("MC: 1.13")) {
 			Utilities.debug("Loading files for version 1.13.1");
 			Bukkit.getServer().getPluginManager().registerEvents(new Listeners1131(this), this);
+            Bukkit.getServer().getPluginManager().registerEvents(new Listeners(this), this);
 		} else if (mcVer.contains("MC: 1.12")) {
 			Utilities.debug("Loading files for version 1.12.2");
-			Bukkit.getServer().getPluginManager().registerEvents(new Listeners188(this), this);
+			Bukkit.getServer().getPluginManager().registerEvents(new Listeners(this), this);
 		} else if (mcVer.contains("MC: 1.8")) {
 			Utilities.debug("Loading files for version 1.8.8");
-			Bukkit.getServer().getPluginManager().registerEvents(new Listeners188(this), this);
+			Bukkit.getServer().getPluginManager().registerEvents(new Listeners(this), this);
 		} else {
 			Utilities.debug("No matching version found.");
 			Utilities.sendConsole(ChatColor.RED + "DISABLED: Server version not supported.");
